Moved sync button to header
